mythdate: match type with Qt, preventing potential Y2106 bug

+++ mythtv/libs/libmythbase/mythdate.cpp
@@ -70,7 +70,7 @@ MBASE_PUBLIC QDateTime fromString(const QString &str, const QString &format)
  *                  at Jan 1 1970 at 00:00:00.
  *  \return A QDateTime.
  */
-MBASE_PUBLIC QDateTime fromSecsSinceEpoch(uint seconds)
+MBASE_PUBLIC QDateTime fromSecsSinceEpoch(int64_t seconds)
 {
     QDateTime dt = QDateTime::fromSecsSinceEpoch(seconds);
     return dt.toUTC();
